feat : add username on header, and split table

+++ routes/users.js
@@ -25,15 +25,15 @@ router.post(
    }),
    (req, res) => {
       if (!req.body.rememberMe) {
-         return res.send({ authenticated: true });
+         return res.redirect("/users/api/user");
       }
       var token = saveRememberMe(req.body.uid);
       res.cookie("remember_me", token, {
          httpOnly: true,
-         maxAge: 60560000,
+         maxAge: 60512312360000,
       });
 
-      return res.send({ authenticated: true });
+      return res.redirect("/users/api/user");
    }
 );
 router.get("/api/login-fail", (req, res) => {
@@ -48,34 +48,38 @@ router.get("/api/login-fail", (req, res) => {
 });
 
 router.get("/api/remember-me", (req, res) => {
-   db.query(
-      `SELECT uid FROM login_token WHERE token="${req.cookies["remember_me"]}";`,
-      (err, row) => {
-         res.clearCookie("remember_me");
-         db.query(
-            `DELETE FROM login_token WHERE token="${req.cookies["remember_me"]}";`
-         );
+   if (req.cookies["remember_me"]) {
+      db.query(
+         `SELECT uid FROM login_token WHERE token="${req.cookies["remember_me"]}";`,
+         (err, row) => {
+            res.clearCookie("remember_me");
+            db.query(
+               `DELETE FROM login_token WHERE token="${req.cookies["remember_me"]}";`
+            );
 
-         if (err || row.length === 0) {
-            res.writeHead(302, { authenticated: false });
-            return res.end();
-         }
-
-         var uid = row[0].uid;
-         var token = saveRememberMe(uid);
-         res.cookie("remember_me", token, {
-            httpOnly: true,
-            maxAge: 60560000,
-         });
-
-         req.login({ UID: uid }, function (err) {
-            if (err) {
-               return res.send({ authenticated: false });
+            if (err || row.length === 0) {
+               res.writeHead(302, { authenticated: false });
+               return res.end();
             }
-            return res.send({ authenticated: true });
-         });
-      }
-   );
+
+            var uid = row[0].uid;
+            var token = saveRememberMe(uid);
+            res.cookie("remember_me", token, {
+               httpOnly: true,
+               maxAge: 60560002131230,
+            });
+
+            req.login({ UID: uid }, function (err) {
+               if (err) {
+                  return res.send({ authenticated: false });
+               }
+               return res.redirect("/users/api/user");
+            });
+         }
+      );
+   } else {
+      return res.send({ authenticated: false });
+   }
 });
 
 router.post("/api/join", (req, res) => {
@@ -83,9 +87,8 @@ router.post("/api/join", (req, res) => {
       .createHmac("sha256", process.env.SHA256_KEY)
       .update(req.body.pw)
       .digest("hex");
-
    db.query(
-      `INSERT INTO auth (UID, NAME, PASSWORD) VALUES ("${req.body.uid}", "${req.body.name}", "${password}");`,
+      `INSERT INTO auth (UID, PASSWORD) VALUES ("${req.body.uid}", "${password}");`,
       (err, ok) => {
          if (err) {
             if (err.code === "ER_DUP_ENTRY") {
@@ -94,7 +97,12 @@ router.post("/api/join", (req, res) => {
                res.send({ isJoined: false });
             }
          } else {
-            res.send({ isJoined: true });
+            db.query(
+               `INSERT INTO userinfo (uid, username) VALUES ("${req.body.uid}", "${req.body.name}");`,
+               (err, ok) => {
+                  res.send({ isJoined: true });
+               }
+            );
          }
       }
    );
@@ -113,7 +121,17 @@ router.post("/api/logout", (req, res) => {
 
 router.get("/api/user", (req, res) => {
    if (req.isAuthenticated()) {
-      res.send({ authenticated: true });
+      db.query(
+         `SELECT username FROM userinfo WHERE uid="${req.session.passport.user}";`,
+         (err, row) => {
+            if (err || !row[0]) {
+               res.send({ authenticated: true });
+            }
+            if (row[0]) {
+               res.send({ authenticated: true, username: row[0].username });
+            }
+         }
+      );
    } else {
       res.send({ authenticated: false });
    }
